+++ subprojects/core/src/main/groovy/org/gradle/cache/internal/filelock/Version1LockStateSerializer.java
@@ -27,7 +27,7 @@ public class Version1LockStateSerializer implements LockStateSerializer {
         return 1;
     }
 
-    public int getVersion() {
+    public byte getVersion() {
         return 1;
     }
